+++ test/binary_sensor_test.py
@@ -222,3 +222,5 @@ class TestBinarySensor(unittest.TestCase):
         xknx = XKNX(loop=self.loop)
         binary_sensor = BinarySensor(xknx, 'TestInput', group_address='1/2/3', group_address_state='1/2/4')
         self.assertEqual(binary_sensor.state_addresses(), [GroupAddress('1/2/4')])
+        binary_sensor2 = BinarySensor(xknx, 'TestInput', group_address='1/2/3')
+        self.assertEqual(binary_sensor2.state_addresses(), [])