+++ xknx/devices/binary_sensor.py
@@ -100,7 +100,9 @@ class BinarySensor(Device):
 
     def state_addresses(self):
         """Return group addresses which should be requested to sync state."""
-        return [self.group_address_state, ]
+        if self.group_address_state is not None:
+            return [self.group_address_state, ]
+        return []
 
     async def _set_internal_state(self, state):
         """Set the internal state of the device. If state was changed after update hooks and connected Actions are executed."""